+++ src/Main.kt
@@ -2,6 +2,6 @@ import problem.*
 
 fun main() {
     println("Welcome to kopt")
-    val problem = Creator[BinPacking]?.invoke("basic", 1)
+    val problem = Creator[Interval]?.invoke("weighted", 1)
     println(problem)
 }