+++ src/discrete/Domain.kt
@@ -8,3 +8,9 @@ class Variables {
 		fun <T> from(items: List<T>): List<Variable> = (0 until items.size).toList()
 	}
 }
+
+class Domain {
+	companion object {
+		fun boolean(): List<Value> = listOf(1, 0)
+	}
+}