+++ src/fn/String.kt
@@ -5,9 +5,15 @@ import discrete.SolutionStringFn
 import discrete.Value
 
 fun <T> stringPartition(values: List<Value>, items: List<T>): SolutionStringFn {
-	fun fn(solution: Solution): String {
+	return fun (solution: Solution): String {
 		val groups = solution.partitionStrings(values, items)
 		return sortedPartitionGroups(groups).joinToString(" ")
 	}
-	return ::fn
+}
+
+fun <T: Comparable<T>> stringSubset(items: List<T>): SolutionStringFn {
+	return fun(solution: Solution): String {
+		val subset = solution.asSubset().mapList(items).sorted()
+		return subset.map { it.toString() }.wrapBraces()
+	}
 }