+++ src/problem/Problems.kt
@@ -4,10 +4,10 @@ import discrete.Problem
 
 const val BinCover = "bin_cover"
 const val BinPacking = "bin_packing"
+const val Interval = "interval"
 
 val Creator: Map<String, (String, Int) -> Problem?>  = mapOf(
 	BinCover to ::newBinCover,
 	BinPacking to ::newBinPacking,
+	Interval to ::newInterval,
 )
-
-fun newName(problem: String, variant: String, n: Int) = "$problem.$variant.$n"